tutorial/requirements.txt
Django==1.10.2
djangorestframework==3.5.1
docutils==0.12
jmespath==0.9.0
pyasn1==0.1.9
Pygments==2.1.3
python-dateutil==2.5.3
rsa==3.4.2
s3transfer==0.1.9
six==1.10.0
mysqlclient==1.3.9